Spring Data JPA [#505127]

+++ src/main/java/ru/job4j/socialmediaapi/repository/ActivityRepository.java
@@ -3,5 +3,5 @@ package ru.job4j.socialmediaapi.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.job4j.socialmediaapi.model.Activity;
 
-public interface ActivityRepository extends CrudRepository<Activity, Long> {
+public interface ActivityRepository extends CrudRepository<Activity, Integer> {
 }
