+++ src/main/java/ru/job4j/socialmediaapi/repository/ChatRepository.java
@@ -3,5 +3,5 @@ package ru.job4j.socialmediaapi.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.job4j.socialmediaapi.model.Chat;
 
-public interface ChatRepository extends CrudRepository<Chat, Long> {
+public interface ChatRepository extends CrudRepository<Chat, Integer> {
 }
