+++ src/main/java/ru/job4j/socialmediaapi/repository/FriendshipRepository.java
@@ -3,5 +3,5 @@ package ru.job4j.socialmediaapi.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.job4j.socialmediaapi.model.Friendship;
 
-public interface FriendshipRepository extends CrudRepository<Friendship, Long> {
+public interface FriendshipRepository extends CrudRepository<Friendship, Integer> {
 }
