+++ src/main/java/ru/job4j/socialmediaapi/repository/ImageRepository.java
@@ -3,5 +3,5 @@ package ru.job4j.socialmediaapi.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.job4j.socialmediaapi.model.Image;
 
-public interface ImageRepository extends CrudRepository<Image, Long> {
+public interface ImageRepository extends CrudRepository<Image, Integer> {
 }
