+++ src/main/java/ru/job4j/socialmediaapi/repository/PostRepository.java
@@ -3,5 +3,5 @@ package ru.job4j.socialmediaapi.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.job4j.socialmediaapi.model.Post;
 
-public interface PostRepository extends CrudRepository<Post, Long> {
+public interface PostRepository extends CrudRepository<Post, Integer> {
 }
