+++ src/main/java/ru/job4j/socialmediaapi/repository/SubscriptionRepository.java
@@ -3,5 +3,5 @@ package ru.job4j.socialmediaapi.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.job4j.socialmediaapi.model.Subscription;
 
-public interface SubscriptionRepository extends CrudRepository<Subscription, Long> {
+public interface SubscriptionRepository extends CrudRepository<Subscription, Integer> {
 }
